REF redo modules to be a bit more sensible
