+++ ngmix/__init__.py
@@ -47,7 +47,7 @@ from . import em
 from . import admom
 from . import gaussmom
 from . import ksigmamom
-from . import prepsfgaussmom
+from . import prepsfmom
 from . import observation
 from .observation import Observation, ObsList, MultiBandObsList
